[SKP-141] refac : styles 파일 분리
