+++ src/screens/Analyze/Analyze.tsx
@@ -1,7 +1,6 @@
-import {Image, StyleSheet, Text, View} from 'react-native';
+import styles from './Analyze.styles';
+import {Image, Text} from 'react-native';
 import {SafeAreaView} from 'react-native-safe-area-context';
-import {theme} from '../../styles';
-import {flexBox, wrapper} from '../../styles/common';
 import {StackScreenProps} from '../../navigators/types';
 import {usePostLocation} from '../../hooks/mutations/location/usePostLocation';
 import {useEffect} from 'react';
@@ -62,27 +61,3 @@ export default function Analyze({navigation, route}: AnalyzeProps) {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    ...wrapper,
-    ...flexBox('column'),
-  },
-  image: {
-    width: 150,
-    height: 150,
-    backgroundColor: '#D9D9D9',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  title: {
-    ...theme.typography.title_sb_21,
-    color: theme.palette.black,
-    marginTop: 50,
-  },
-  subtitle: {
-    ...theme.typography.title_m_16,
-    color: theme.palette.gray5,
-    marginTop: 11,
-  },
-});
